deleted sample.js. use npm run seed. working but no associations
