+++ package.json
@@ -17,7 +17,6 @@
 		"prettify": "prettier --write \"**/*.{js,jsx,json,css,scss,md}\"",
 		"postinstall": "touch secrets.js",
 		"seed": "node script/seed.js",
-		"sampleSeed": "node script/sample.js",
 		"start": "node server",
 		"start-dev": "NODE_ENV='development' npm run build-client-watch & NODE_ENV='development' npm run start-server",
 		"start-server": "nodemon server -e html,js,scss --ignore public --ignore client",
